Track number of entries with max PSL.
Although only 1 entry can be inserted at a time, and insertions are
blocked once any entry hits the PSL threshold, it is possible for more
than one entry to hit the PSL limit as a result of multiple shifts
during an insertion.  (E.g. consecutive bucket groups that are just
below the limit could both be shifted one position farther away from
their ideal positions.)

Thus, a boolean flag is not sufficient to track the "at PSL limit"
state of the table.  It is necessary to track the *number* of entries
that are at the maximum PSL.

* Rename sht_ht.max_psl to sht_ht.biggest_psl to make its purpose
  somewhat clearer.

* Replace the sht_ht.psl_maxxed flag with sht_ht.max_psl_ct, which
  counts the number of entries at the maximum PSL.

* Move wrap-around logic from sht_remove_at() to sht_shift_wrap()
  helper function.